+++ .vscode/settings.json
@@ -14,6 +14,7 @@
         "fromlist",
         "hyperparameters",
         "hyperparams",
+        "imagenet",
         "ipython",
         "itwinai",
         "Lockfiles",
