--- tutorials/distributed-ml/torch-scaling-test/ddp/config.yaml
@@ -1,23 +1,0 @@
-# I/O
-data_dir: /p/project/intertwin/datasets/Imagenet_sub/ImageNet_uncompressed/train/ #/p/project/intertwin/datasets/ImageNet_uncompressed/train
-restart_int: 10
-verbose: True
-
-# Model
-batch_size: 64
-epochs: 3
-lr: 0.001
-momentum: 0.5
-shuff: False
-
-# Debugging
-benchrun: False
-testrun: False
-nseed: 10
-log_int: 10
-
-# Distributed ML
-backend: nccl
-nworker: 4 # num workers dataloader
-prefetch: 2
-no_cuda: False
