--- tutorials/distributed-ml/torch-scaling-test/deepspeed/config.yaml
@@ -1,17 +1,0 @@
-# I/O
-data_dir: /p/project/intertwin/datasets/Imagenet_sub/ImageNet_uncompressed/train/ #/p/project/intertwin/datasets/ImageNet_uncompressed/train
-restart_int: 10
-verbose: True
-
-# Model
-batch_size: 64
-epochs: 3
-lr: 0.001
-
-# Debugging
-testrun: False
-log_int: 10
-
-# Distributed ML
-backend: nccl
-nworker: 4 # num workers dataloader
