--- tutorials/distributed-ml/torch-scaling-test/horovod/config.yaml
@@ -1,20 +1,0 @@
-# I/O
-data_dir: /p/project/intertwin/datasets/Imagenet_sub/ImageNet_uncompressed/train/ #/p/project/intertwin/datasets/ImageNet_uncompressed/train
-verbose: True
-
-# Model
-batch_size: 64
-epochs: 3
-lr: 0.001
-momentum: 0.5
-use_adasum: False
-
-# Debugging
-seed: 10
-log_interval: 10
-
-# Distributed ML
-nworker: 4 # num workers dataloader
-no_cuda: False
-fp16_allreduce: False
-gradient_predivide_factor: 1.0
