+++ tutorials/distributed-ml/torch-scaling-test/hvd_slurm.sh
@@ -1,19 +1,19 @@
 #!/bin/bash
 
 # general configuration of the job
-#SBATCH --job-name=Torch_DeepSpeed_tutorial-1
+#SBATCH --job-name=Torch_HVD_tutorial-1
 #SBATCH --account=intertwin
 #SBATCH --mail-user=
 #SBATCH --mail-type=ALL
-#SBATCH --output=job-ds.out
-#SBATCH --error=job-ds.err
+#SBATCH --output=job-hvd.out
+#SBATCH --error=job-hvd.err
 #SBATCH --time=00:30:00
 
 # configure node and process count on the CM
 #SBATCH --partition=batch
 #SBATCH --nodes=2
 #SBATCH --ntasks-per-node=4
-#SBATCH --cpus-per-task=4
+#SBATCH --cpus-per-task=8
 #SBATCH --gpus-per-node=4
 #SBATCH --exclusive
 
@@ -24,7 +24,7 @@
 ml Stages/2024 GCC OpenMPI CUDA/12 MPI-settings/CUDA Python HDF5 PnetCDF libaio mpi4py
 
 # set env
-source ../../../../envAI_hdfml/bin/activate
+source ../../../envAI_hdfml/bin/activate
 
 # job info
 debug=false
@@ -44,7 +44,8 @@ if [ "$debug" = true ] ; then
 fi
 echo
 
-# set env vars
+# set vars
+# export NCCL_DEBUG=INFO
 export SRUN_CPUS_PER_TASK=${SLURM_CPUS_PER_TASK}
 export OMP_NUM_THREADS=1
 if [ "$SLURM_CPUS_PER_TASK" -gt 0 ] ; then
@@ -53,22 +54,7 @@ fi
 export CUDA_VISIBLE_DEVICES="0,1,2,3"
 
 # launch training
-MASTER_ADDR=$(scontrol show hostnames "\$SLURM_JOB_NODELIST" | head -n 1)i
-export MASTER_ADDR
-export MASTER_PORT=29500 
+TRAINING_CMD="horovod_trainer.py -c horovod-config.yaml"
 
-TRAINING_CMD="DS_trainer.py -c config.yaml"
-
-# Run without launcher: set --ntasks-per-node=NUM_GPUS
-srun --cpu-bind=none python -u $TRAINING_CMD --deepspeed
-
-# # Run with deepspeed launcher: set --ntasks-per-node=1
-# # https://www.deepspeed.ai/getting-started/#multi-node-environment-variables
-# export NCCL_IB_DISABLE=1
-# export NCCL_SOCKET_IFNAME=eth0
-# nodelist=$(scontrol show hostname $SLURM_NODELIST)
-# echo "$nodelist" | sed -e 's/$/ slots=4/' > .hostfile
-# # Requires passwordless SSH access among compute node
-# srun --cpu-bind=none deepspeed --hostfile=.hostfile $TRAINING_CMD --deepspeed
-# rm .hostfile
+srun --cpu-bind=none python -u $TRAINING_CMD
