+++ tutorials/distributed-ml/torch-scaling-test/runall.sh
@@ -1,15 +1,9 @@
 #!/bin/bash
 # Run all versions of distributed ML version
 
-for fold in ddp horovod deepspeed
+for name in ddp hvd deepspeed
 do
-    cd $fold
     rm *checkpoint.pth.tar *.out *.err *.csv
-    if [ $fold == "horovod" ]
-    then
-        fold="hvd"
-    fi
     # echo $fold" training: $(sbatch --nodes=1 $fold"_slurm.sh")"
-    echo $fold" training: $(sbatch $fold"_slurm.sh")"
-    cd ..
+    echo $name" training: $(sbatch $name"_slurm.sh")"
 done
